requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
frozenlist==1.6.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
